requirements.txt
#pip install -r requirements.txt <-- rodar no terminal
mysql-connector-python